+++ src/fparser/two/Fortran2008/Fortran2008.py
@@ -143,449 +143,6 @@ from fparser.two.Fortran2003 import (
 )
 
 
-class Data_Component_Def_Stmt(Data_Component_Def_Stmt_2003):  # R436
-    """
-    Fortran 2008 rule 436.
-
-    .. code-block:: fortran
-
-        data-component-def-stmt is declaration-type-spec [
-                 [ , component-attr-spec-list ] :: ] component-decl-list
-
-    The implementation of this rule does not add anything to the Fortran 2003
-    variant but reimplements the match method identical to Fortran 2003 as
-    otherwise the generated Fortran 2008 variant of `Component_Attr_Spec_List`
-    would not be used. Unfortunately, the required `attr_spec_list_cls` can not
-    simply be provided as a class property since the relevant class is only
-    generated at the end of this file using the `use_names` class property of
-    this class.
-
-    Associated constraints are:
-
-    "C439 (R436)  No component-attr-spec shall appear more than once in a given
-          component-def-stmt."
-    "C440 (R436)  If neither the POINTER nor the ALLOCATABLE attribute is
-          specified, the declaration-type-spec in the component-def-stmt shall
-          specify an intrinsic type or a previously defined derived type."
-    "C441 (R436)  If the POINTER or ALLOCATABLE attribute is specified, each
-          component-array-spec shall be a deferred-shape-spec-list."
-    "C442 (R436)  If a coarray-spec appears, it shall be a
-          deferred-coshape-spec-list and the component shall have the
-          ALLOCATABLE attribute."
-    "C443 (R436)  If a coarray-spec appears, the component shall not be of type
-          C_PTR or C_FUNPTR."
-    "C445 (R436)  If neither the POINTER nor the ALLOCATABLE attribute is
-          specified, each component-array-spec shall be an
-          explicit-shape-spec-list."
-    "C447 (R436)  A component shall not have both the ALLOCATABLE and POINTER
-          attributes."
-    "C448 (R436)  If the CONTIGUOUS attribute is specified, the component shall
-          be an array with the POINTER attribute."
-    "C457 (R436)  If component-initialization appears, a double-colon separator
-          shall appear before the component-decl-list."
-    "C458 (R436)  If component-initialization appears, every type parameter and
-          array bound of the component shall be a colon or constant expression.
-    "C459 (R436)  If => appears in component-initialization, POINTER shall
-          appear in the component-attr-spec-list. If = appears in
-          component-initialization, neither POINTER nor ALLOCATABLE shall
-          appear in the component-attr-spec-list."
-
-    C439-C443, C445, C447-C448, C457-C459 are currently not checked
-    - issue #258.
-
-    """
-
-    @staticmethod
-    def match(string):
-        """Implements the matching of a data component definition statement.
-
-        :param str string: the reader or string to match as a data \
-                           component definition statement.
-
-        :return: a 3-tuple containing declaration type specification, \
-                 component attribute specification and component declaration \
-                 list if there is a match or None if there is no match.
-        :rtype: `NoneType` or \
-            (:py:class:`fparser.two.Fortran2003.Declaration_Type_Spec`, \
-             :py:class:`fparser.two.Fortran2008.Component_Attr_Spec_List`, \
-             :py:class:`fparser.two.Fortran2003.Component_Decl_List`)
-
-        """
-        # Avoid circular dependencies by importing here.
-        # pylint: disable=import-outside-toplevel
-        from fparser.two.Fortran2008 import Component_Attr_Spec_List
-
-        return Type_Declaration_StmtBase.match(
-            Declaration_Type_Spec, Component_Attr_Spec_List, Component_Decl_List, string
-        )
-
-
-class Component_Attr_Spec(Component_Attr_Spec_2003):  # R437
-    """
-    Fortran 2008 rule R437.
-
-    .. code-block:: fortran
-
-        component-attr-spec is access-spec
-                               or ALLOCATABLE
-                               or CODIMENSION lbracket coarray-spec rbracket
-                               or CONTIGUOUS
-                               or DIMENSION ( component-array-spec )
-                               or POINTER
-
-    In the spec above, lbracket and rbracket are left and right square
-    brackets `[]` but not printed explicitly to avoid misinterpretation
-    as optional parts.
-
-    This rule adds CODIMENSION and CONTIGUOUS attributes to Fortran2003's R441.
-
-    """
-
-    subclass_names = Component_Attr_Spec_2003.subclass_names[:]
-    subclass_names.append("Codimension_Attr_Spec")
-    attributes = Component_Attr_Spec_2003.attributes[:]
-    attributes.append("CONTIGUOUS")
-
-
-class Type_Declaration_Stmt(Type_Declaration_Stmt_2003):  # R501
-    """
-    Fortran 2008 rule 501.
-
-    .. code-block:: fortran
-
-        type-declaration-stmt is declaration-type-spec [ [ , attr-spec ] ... :: ]
-                                 entity-decl-list
-
-    The implementation of this rule does not add anything to the Fortran 2003
-    variant but overwrites :py:meth:`get_attr_spec_list_cls` to use
-    the Fortran 2008 variant of :py:class:`Attr_Spec_List`.
-
-    Associated constraints are:
-
-    "C501 (R501)  The same attr-spec shall not appear more than once in a given
-          type-declaration-stmt."
-    "C502 (R501)  If a language-binding-spec with a NAME= specifier appears,
-          the entity-decl-list shall consist of a single entity-decl."
-    "C503 (R501)  If a language-binding-spec is specified, the entity-decl-list
-          shall not contain any procedure names."
-    "C505 (R501)  If initialization appears, a double-colon separator shall
-          appear before the entity-decl-list."
-
-    C501-C503, C505 are currently not checked - issue #259.
-
-    """
-
-    @staticmethod
-    def get_attr_spec_list_cls():
-        """Return the type used to match the attr-spec-list
-
-        This overwrites the Fortran 2003 type with the Fortran 2008 variant.
-
-        """
-        # Avoid circular dependencies by importing here.
-        # pylint: disable=import-outside-toplevel
-        from fparser.two.Fortran2008 import Attr_Spec_List
-
-        return Attr_Spec_List
-
-
-class Codimension_Attr_Spec(WORDClsBase):  # R502.d
-    """
-    codimension-attr-spec is CODIMENSION lbracket coarray-spec rbracket
-
-    In the spec above, lbracket and rbracket are left and right square
-    brackets `[]` but not printed explicitly to avoid misinterpretation
-    as optional parts.
-
-    """
-
-    subclass_names = []
-    use_names = ["Coarray_Bracket_Spec"]
-
-    @staticmethod
-    def match(string):
-        """
-        Implements the matching for the CODIMENSION attribute.
-
-        :param str string: the string to match as the attribute.
-
-        :return: `None` if there is no match, otherwise a 2-tuple \
-                 containing `CODIMENSION` as a string and the matched \
-                 coarray-spec..
-        :rtype: `NoneType` or \
-            (`str`, :py:class:`fparser.two.Fortran2008.Coarray_Bracket_Spec`,)
-
-        """
-        return WORDClsBase.match(
-            "CODIMENSION", Coarray_Bracket_Spec, string, colons=False, require_cls=True
-        )
-
-
-class Coarray_Bracket_Spec(BracketBase):  # R502.d.0
-    """
-    coarray-bracket-spec is lbracket coarray-spec rbracket
-
-    In the spec above, lbracket and rbracket are left and right square
-    brackets `[]` but not printed explicitly to avoid misinterpretation
-    as optional parts.
-
-    """
-
-    subclass_names = []
-    use_names = ["Coarray_Spec"]
-
-    @staticmethod
-    def match(string):
-        """
-        Implements the matching for the coarray specification
-        including the square brackets.
-
-        :param str string: the string to match as the specification.
-
-        :return: `None` if there is no match, otherwise a 3-tuple \
-                 containing the left bracket, the matched coarray-spec, \
-                 and the right bracket.
-        :rtype: `NoneType` or \
-            (`str`, :py:class:`fparser.two.Fortran2008.Coarray_Spec`, `str`)
-
-        """
-        return BracketBase.match("[]", Coarray_Spec, string)
-
-
-class Attr_Spec(Attr_Spec_2003):  # R502
-    """
-    Fortran 2008 rule R502.
-
-    .. code-block:: fortran
-
-        attr-spec is access-spec
-                     or ALLOCATABLE
-                     or ASYNCHRONOUS
-                     or CODIMENSION lbracket coarray-spec rbracket
-                     or CONTIGUOUS
-                     or DIMENSION ( array-spec )
-                     or EXTERNAL
-                     or INTENT ( intent-spec )
-                     or INTRINSIC
-                     or language-binding-spec
-                     or OPTIONAL
-                     or PARAMETER
-                     or POINTER
-                     or PROTECTED
-                     or SAVE
-                     or TARGET
-                     or VALUE
-                     or VOLATILE
-
-    In the spec above, lbracket and rbracket are left and right square
-    brackets `[]` but not printed explicitly to avoid misinterpretation
-    as optional parts.
-
-    This rule adds CODIMENSION and CONTIGUOUS attributes to Fortran2003's R503.
-
-    """
-
-    subclass_names = Attr_Spec_2003.subclass_names[:]
-    subclass_names.append("Codimension_Attr_Spec")
-    use_names = []
-
-    @staticmethod
-    def match(string):
-        """
-        Implements the matching for attributes of types.
-
-        :param str string: the string to match as attribute.
-
-        :return: `None` if there is no match, otherwise a 1-tuple \
-                 containing the matched string.
-        :rtype: `NoneType` or (`str`,)
-
-        """
-        return STRINGBase.match(pattern.abs_attr_spec_f08, string)
-
-
-class Coarray_Spec(Base):  # R509
-    """
-    Fortran 2008 rule R509.
-
-    .. code-block:: fortran
-
-        coarray-spec is deferred-coshape-spec-list
-                        or explicit-coshape-spec-list
-
-    """
-
-    subclass_names = ["Explicit_Coshape_Spec", "Deferred_Coshape_Spec_List"]
-
-
-class Deferred_Coshape_Spec(SeparatorBase):  # R510
-    """
-    Fortran 2008 rule R510
-    deferred-coshape-spec is :
-
-    """
-
-    subclass_names = []
-
-    @staticmethod
-    def match(string):
-        """
-        Implements the matching for deferred coarray shape specification.
-
-        :param str string: the string to match as deferred shape.
-
-        :return: `None` if there is no match, otherwise a 2-tuple \
-                 containing `None`.
-        :rtype: `NoneType` or (`None`, `None`)
-
-        """
-        if string == ":":
-            return (None, None)
-        return None
-
-
-class Explicit_Coshape_Spec(SeparatorBase):  # R511
-    """
-    Fortran 2008 rule R511
-    explicit-coshape-spec is [ coshape-spec-list , ] [ lower-cobound : ] *
-
-    Associated constraint is:
-
-    "C529 (R511)  A lower-cobound or upper-cobound that  is  not  a  constant
-          expression shall appear only in a subprogram, BLOCK construct, or
-          interface body."
-
-    C529 is currently not checked - issue #259.
-
-    """
-
-    subclass_names = []
-    use_names = ["Coshape_Spec_List", "Lower_Cobound"]
-
-    @staticmethod
-    def match(string):
-        """
-        Implements the matching for explicit coarray shape specification.
-
-        :param str string: the string to match as deferred shape.
-
-        :return: `None` if there is no match, otherwise a 2-tuple \
-                 containing matched coshape-spec-list or `None` and \
-                 matched lower-cobound or `None`.
-        :rtype: `NoneType` or \
-            (:py:class:`fparser.two.Fortran2008.Coshape_Spec_List` or `None`, \
-             :py:class:`fparser.two:Fortran2008.Lower_Cobound` or `None`)
-
-        """
-        # Avoid circular dependencies by importing here.
-        # pylint: disable=import-outside-toplevel
-        from fparser.two.Fortran2008 import Coshape_Spec_List
-
-        if not string.endswith("*"):
-            return None
-        line = string[:-1].rstrip()
-        if not line:
-            return (None, None)
-        if line.endswith(":"):
-            line, repmap = string_replace_map(line[:-1].rstrip())
-            sep_pos = line.rfind(",")
-            if sep_pos == -1:
-                return (None, Lower_Cobound(repmap(line)))
-            return (
-                Coshape_Spec_List(repmap(line[:sep_pos].rstrip())),
-                Lower_Cobound(repmap(line[sep_pos + 1 :].lstrip())),
-            )
-        if not line.endswith(","):
-            return None
-        line = line[:-1].rstrip()
-        return (Coshape_Spec_List(line), None)
-
-    def tostr(self):
-        """
-        Converts the explicit coarray shape specification to string.
-
-        :return: the shape specification as string.
-        :rtype: str
-
-        """
-        s = ""
-        if self.items[0]:
-            s += str(self.items[0]) + ", "
-        if self.items[1]:
-            s += str(self.items[1]) + " : "
-        s += "*"
-        return s
-
-
-class Coshape_Spec(SeparatorBase):  # R511.a
-    """
-    coshape-spec is [ lower-cobound : ] upper-cobound
-
-    """
-
-    subclass_names = []
-    use_names = ["Lower_Cobound", "Upper_Cobound"]
-
-    @staticmethod
-    def match(string):
-        """
-        Implements the matching for a coarray shape.
-
-        :param str string: the string to match as shape.
-
-        :return: `None` if there is no match, otherwise a 2-tuple with \
-                 lower bound if given or `None`, and upper bound.
-        :rtype: `NoneType` or \
-            (`None`, :py:class:`fparser.two.Fortran2008.Upper_Cobound`) or \
-            (:py:class:`fparser.two.Fortran2008.Lower_Cobound`, \
-             :py:class:`fparser.two.Fortran2008.Upper_Cobound`)
-
-        """
-        line, repmap = string_replace_map(string)
-        if ":" not in line:
-            return (None, Upper_Cobound(string))
-        lower, upper = line.split(":", 1)
-        lower = lower.rstrip()
-        upper = upper.lstrip()
-        if not upper:
-            return None
-        if not lower:
-            return None
-        return (Lower_Cobound(repmap(lower)), Upper_Cobound(repmap(upper)))
-
-    def tostr(self):
-        """
-        Converts the Shape specification to string.
-
-        :return: the shape specification as string.
-        :rtype: str
-
-        """
-        if self.items[0] is None:
-            return str(self.items[1])
-        return SeparatorBase.tostr(self)
-
-
-class Lower_Cobound(Base):  # R512
-    """
-    Fortran 2008 rule R512
-    lower-cobound is specification-expr
-
-    """
-
-    subclass_names = ["Specification_Expr"]
-
-
-class Upper_Cobound(Base):  # R513
-    """
-    Fortran 2008 rule R513
-    upper-cobound is specification-expr
-
-    """
-
-    subclass_names = ["Specification_Expr"]
-
-
 class Do_Term_Action_Stmt(Do_Term_Action_Stmt_2003):  # R826
     """
     Fortran 2008 rule R826.
